* src/bar.c(_error_write): Fix another potential race condition
In this case, the lock acquisition and release is done around the for
loop instead instead of around the call to wget_bar_update. By creating
such a coarse critical section, we ensure that the entire screen has
been redrawn for the user before any further changes to it are made.
This is especially the case when error_printf() is invoked in quick
succession which would cause multiple instances of wget_update_thread()
to run simultaneously, greatly increasing the lock contention.