Clean up progress bar code
Saving the max, cols and ratio to the slot structure was an optimization
used to prevent unneeded refreshes of the progress bar on slow
connections. However, with the asynchronous progress bar, this
optimization is no longer necessary as we refresh the progress bar far
fewer times already. Also, as other elements like the ETA are added to
the progress bar, these checks will only cause more problems as the ETA
would need to be updated even if no data has been downloaded in some
time.

* libwget/bar.c: Remove ratio, max, cur, cols and first from the
_wget_slot_st struct and eliminate all references to them.