* src/stats.c: Initialize all stats modules, always
This is a terrible hack. But it's the smallest solution I can come up
with, with the existing stats code to fix the problem that none of the
stats are being reported.

The problem is that stats_init() must be called before the command line
is parsed. But stats_init() depends on the command line parsing code to
set the status flags for which stats to report. Hence, it always sees
them all as negative.

This commit will initialize all the stats modules, always. So, the stats
are always collected. Only, we decide at the very end, what, if
anything, to display. As is obvious, this is terrible. It's a waste of
resources to compute all these stats that are simply not required. But,
till the stats code is re-written cleanly, this is what we have to deal
with