Add new option --keep-extension
* src/options.c: Add option --keep-extension
* src/wget.c (_open_unique): Add code for --keep-extension
* src/wget_options.h (struct config): Add keep_extension member
* docs/wget2.md: Add docs for --keep-extension

The new option --keep-extension changes the behavior for creating a
unique filename if a file already exists.
The standard (default) pattern is to generate <filename>.<N>, the new
pattern is <basename>_<N>.<ext>.

The idea is use such files without renaming when the use depends on the
extension, like on Windows.