Ensure that robots.txt is honoured even with -c
There was a bug that prevented the robots.txt from being parsed when
resuming an incomplete download. This would cause the robots rules to
not be honoured when resuming a broken download.

* src/wget.c (process_response): Move the robots check out of the
conditional, so it runs in all valid cases
(http_create_request): In case -c has been passed and we see an existing
robots.txt, remove it. As a result, we will always download a fresh
copy of the robots.txt file
* tests/test-robots.c: Add a new test to ensure the robots file is
always being parsed