Support XDG Desktop Specification for config files
Adds functionality to Wget2 to support the XDG Base Directory
Specification[1]. Also marks as deprecated the old location of the
config file.

* docs/wget2.md: Explicitly mark `$HOME/.wget2rc` as deprecated.
* po/potfiles.in: utils.c now has a translateable stirng
* src/options.c(read_config): Use bool instead of int
(get_config_files): Now takes config_home and user_home as input.
`user_home` is only temporary and will be removed once support for
~/.wget2rc is removed.
(get_xdg_data_home): Derives the location where the data files are to be
stored according to [1].
(get_xdg_config_home): Derives the location where the config files are
to be stored according to [1].
(init): Store all the data files in $XD_DATA_HOME
* src/wget.c(mkdir_path): Move function to utils.c
* src/utils.c(mkdir_path): Move function from wget.c. Can now handle
  both file and directory paths.
* src/utils.h: Add declaration for mkdir_path

[1]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html