OpenSSL: Fix CRL checking
Check CRLs properly on the OpenSSL backend. CRLs were previously bypassed
because OpenSSL's internal certificate verification function was being overridden.

We need to provide our own verification callback for HPKP and OCSP, but this callback is now
introduced later in the verification chain, so that it does no longer override OpenSSL's checks.

This commit also re-enables a CRL test on OpenSSL, which would previously only run on the GnuTLS backend.

* libwget/ssl_openssl.c (openssl_init): Register verification callback, rather than revocation one.
  (_openssl_revocation_check_fn): Adhere to verification callback interface.
* tests/test-i-https.c: Enable CRL test for OpenSSL (but keep it disabled for wolfSSL)

Signed-off-by: Ander Juaristi <a@juaristi.eus>