Adapt configure.ac to support multiple SSL/TLS backends
This patch modifies `configure.ac` to support GnuTLS, OpenSSL and wolfSSL as the SSL/TLS backends.

Either of these can be enabled at compile time the `configure` switch `--with-ssl`. Example:

    ./configure --with-ssl=openssl

If `--with-ssl` is omitted, the default behavior will be to build with GnuTLS if available, and without
TLS if not. TLS can be explicitly disabled with `--without-ssl`.

Regardless of which specific backend has been selected, the TLS and OCSP tests will always be built with GnuTLS,
as that is the only backend currently supported in the test suite.

It renames the constant `HAVE_GNUTLS_OCSP_H` to a more generic `WITH_OCSP`, that captures
the notion that OCSP has been enabled at compile-time, hiding the specific backend.

Changed files

 * configure.ac: honor --with-ssl option, but keep on testing for GnuTLS
   for the test suite.
 * libwget/ssl_gnutls.c: replace HAVE_GNUTLS_OCSP_H with WITH_OCSP
 * libwget/ssl_openssl.c: likewise
 * tests/libtest.c: replace HAVE_GNUTLS_OCSP_H with WITH_OCSP.
   Check for WITH_GNUTLS_IN_TESTSUITE.