dns: Merge SOCK_STREAM and SOCK_DGRAM addrinfo lists
We want resolve() to return both, TCP and UDP sockets.

On Linux, we could set .ai_socktype = 0 on the getaddrinfo() hints which
would return us all the socket types available. However, that is not
accepted on Windows. Instead, we ask for TCP and UDP sockets separately
and then merge the two lists.